third_party/blink/web_tests/fast/xsl/xslt/xslt-string-parameters-expected.txt

CONSOLE WARNING: Usage of XSLTProcessor or XSLT Processing Instructions was detected. These features have been deprecated by all browsers, and a special early trial of complete removal is underway in this browser.
--> If you are a *user* experiencing a problem, please report the issue directly to the operator of the website.
--> If you are a site owner, and you think this trial is causing an unexpected issue, please report a bug at https://issues.chromium.org/issues/new?component=1456730&template=2210866
CONSOLE ERROR: Uncaught NotSupportedError: Failed to construct 'XSLTProcessor': XSLT is disabled
This tests that passing string parameters to the XSLTProcessor works as expected. If this test is successful, the text "SUCCESS" will be shown below.

FAILURE
